Promoting consistently passing tests to Main suite
